Update README.md
Proposing to add the Mathematical Mesh to the protocols section.

The Mesh is basically a Swiss army knife of cryptography. It supports everything OpenPGP, S/MIME and PKIX do. Signal type capabilities in a non-walled garden environment are being added.

The Mesh potentially allows an entirely end-to-end encrypted social media/collaboration experience. That is with all the documents on the host being encrypted such that the hosting site has no access to the plaintext or the decryption keys.